+++ src/com/engeto/restaurant/DishList.java
@@ -3,6 +3,7 @@ package com.engeto.restaurant;
 *   DishList - List of all possible dishes tha our cooks are able to cook (It is a base for creation of menu)
  */
 import java.io.*;
+import java.math.BigDecimal;
 import java.time.DateTimeException;
 import java.util.ArrayList;
 import java.util.List;
@@ -71,9 +72,8 @@ public class DishList {
                 lineNumber++;
                 line = scanner.nextLine();
                 items = line.split(Settings.DELIMITER);
-                Dish dish = new Dish(items[0], Double.parseDouble(items[1]), Integer.parseInt(items[2]),
+                Dish dish = new Dish(items[0], BigDecimal.valueOf(Double.parseDouble(items[1])), Integer.parseInt(items[2]),
                         items[3], Enum.valueOf(Category.class, items[4]));
-                                // getEnumFormat(Enum.valueOf(Category.class, items[4]))
                 dishList.add(dish);
             }
         } catch (FileNotFoundException e) {
